Finalize some things: remove headers from QueryInput; move ActivityInput;
