+++ src/Interceptor/WorkflowInbound/QueryInput.php
@@ -9,7 +9,6 @@
 
 namespace Temporal\Interceptor\WorkflowInbound;
 
-use Temporal\DataConverter\HeaderInterface;
 use Temporal\DataConverter\ValuesInterface;
 
 class QueryInput
@@ -17,19 +16,15 @@ class QueryInput
     public function __construct(
         public string $queryName,
         public ValuesInterface $arguments,
-        // todo: remove headers
-        public HeaderInterface $header,
     ) {
     }
 
     public function with(
         ValuesInterface $arguments = null,
-        HeaderInterface $header = null,
     ): self {
         return new self(
             $this->queryName,
             $arguments ?? $this->arguments,
-            $header ?? $this->header
         );
     }
 }
