+++ src/Internal/Transport/Router/InvokeActivity.php
@@ -16,8 +16,8 @@ use Temporal\Activity;
 use Temporal\Activity\ActivityInfo;
 use Temporal\DataConverter\EncodedValues;
 use Temporal\Exception\DoNotCompleteOnResultException;
+use Temporal\Interceptor\ActivityInbound\ActivityInput;
 use Temporal\Interceptor\ActivityInboundInterceptor;
-use Temporal\Interceptor\WorkflowClient\ActivityInput;
 use Temporal\Internal\Activity\ActivityContext;
 use Temporal\Internal\Declaration\Prototype\ActivityPrototype;
 use Temporal\Internal\Interceptor\PipelineProvider;
